0x00. Hello, World

+++ 0x00-hello_world/6-size.c
@@ -9,8 +9,8 @@ int main(void)
 {
 	char c;
 	int i;
-	long int j;
-	long long int k;
+	long j;
+	double k;
 	float f;
 
 	printf("Size of a char: %lu byte(s)\n", sizeof(c));
